refactor(test): extract shared test utils into boxlite-test-utils crate
- Move warm cache infrastructure (image pre-pull, symlinked cache dirs,
  warm_temp_dir) from boxlite/tests/common into a shared boxlite-test-utils crate
- CLI tests now use per-test TempDir with symlinked caches (parallel-safe)
- Remove serial test group constraint from nextest config
- Fix registry tests to use TEST_REGISTRIES mirrors instead of docker.io
- Fix fully_qualified test: add working mirror for guest rootfs pull
- Update box ID regex in create tests to match new nanoid format
