+++ boxlite-cli/Cargo.toml
@@ -42,6 +42,7 @@ gtmpl_value = "0.5"
 dirs = "6.0"
 
 [dev-dependencies]
+boxlite-test-utils = { path = "../boxlite-test-utils" }
 assert_cmd = "2.1.1"
 predicates = "3.1.3"
 rstest = "0.21"
