+++ boxlite-cli/tests/common/mod.rs
@@ -1,31 +1,31 @@
 #![allow(dead_code)]
 
 use assert_cmd::Command;
+use boxlite_test_utils::{TEST_REGISTRIES, warm_temp_dir};
 use std::path::PathBuf;
-use std::sync::{Mutex, MutexGuard, OnceLock};
 use std::time::Duration;
+use tempfile::TempDir;
 
-/// images pre-pull (manually maintained to avoid Docker Hub rate limits)
-const TEST_IMAGES: &[&str] = &["alpine:latest", "python:alpine"];
-
-// Prevents "Failed to acquire runtime lock" errors since all tests share the same home dir
-static TEST_LOCK: OnceLock<Mutex<()>> = OnceLock::new();
-static SHARED_HOME: OnceLock<PathBuf> = OnceLock::new();
+fn apply_registries(cmd: &mut Command) {
+    for reg in TEST_REGISTRIES {
+        cmd.arg("--registry").arg(reg);
+    }
+}
 
 pub struct TestContext {
     pub cmd: Command,
-    pub home: &'static PathBuf,
-    // Hold the lock until the test is done
-    pub _guard: MutexGuard<'static, ()>,
+    pub home: PathBuf,
+    _tmp_dir: TempDir,
 }
 
 impl TestContext {
-    /// sharing the same context (home dir and lock)
+    /// Create a new command sharing the same home directory.
     pub fn new_cmd(&self) -> Command {
         let bin_path = env!("CARGO_BIN_EXE_boxlite");
         let mut cmd = Command::new(bin_path);
         cmd.timeout(Duration::from_secs(60));
-        cmd.arg("--home").arg(self.home);
+        cmd.arg("--home").arg(&self.home);
+        apply_registries(&mut cmd);
         cmd
     }
 
@@ -42,58 +42,33 @@ impl TestContext {
     }
 }
 
-pub fn boxlite() -> TestContext {
-    let lock = TEST_LOCK.get_or_init(|| Mutex::new(()));
-    let guard = lock.lock().unwrap_or_else(|e| e.into_inner());
-
-    let home = SHARED_HOME.get_or_init(|| {
-        eprintln!("Initializing shared test environment...");
-
-        // Use a very short path in /tmp to avoid SUN_LEN limits for Unix sockets (104-108 chars)
-        // Project folders can be very deep, exceeding this limit when appended with /boxes/.../sockets/ready.sock
-        let test_home = PathBuf::from("/tmp/bl");
-        std::fs::create_dir_all(&test_home).expect("Failed to create /tmp/bl directory");
-
-        let home = test_home;
-        let bin_path = env!("CARGO_BIN_EXE_boxlite");
-
-        // Clean up any stale containers from previous interrupted test runs (Ctrl+C, kill, etc.)
-        eprintln!("Cleaning up stale containers...");
-        let _ = std::process::Command::new(bin_path)
-            .args(["--home", home.to_str().unwrap(), "rm", "-fa"])
-            .output();
-
-        // Pre-pull test images to avoid Docker Hub rate limits
-        eprintln!("Pre-pulling {} test image(s)...", TEST_IMAGES.len());
-
-        for image in TEST_IMAGES {
-            let result = std::process::Command::new(bin_path)
-                .args(["--home", home.to_str().unwrap(), "pull", image])
-                .output();
-
-            match result {
-                Ok(output) if output.status.success() => {
-                    eprintln!("  ✓ {}", image);
-                }
-                _ => {
-                    eprintln!("  ⚠ {} (will pull on-demand)", image);
-                }
-            }
-        }
-
-        eprintln!("Test environment ready");
-        home
-    });
-
-    let bin_path: &str = env!("CARGO_BIN_EXE_boxlite");
+/// Create a TestContext without default registries.
+/// Use this when the test needs full control over which registries are used.
+pub fn boxlite_bare() -> TestContext {
+    let (tmp_dir, home) = warm_temp_dir();
+    let bin_path = env!("CARGO_BIN_EXE_boxlite");
     let mut cmd = Command::new(bin_path);
-    // You can override this with .timeout(Duration::from_secs(N))
     cmd.timeout(Duration::from_secs(60));
-    cmd.arg("--home").arg(home);
+    cmd.arg("--home").arg(&home);
 
     TestContext {
         cmd,
         home,
-        _guard: guard,
+        _tmp_dir: tmp_dir,
+    }
+}
+
+pub fn boxlite() -> TestContext {
+    let (tmp_dir, home) = warm_temp_dir();
+    let bin_path = env!("CARGO_BIN_EXE_boxlite");
+    let mut cmd = Command::new(bin_path);
+    cmd.timeout(Duration::from_secs(60));
+    cmd.arg("--home").arg(&home);
+    apply_registries(&mut cmd);
+
+    TestContext {
+        cmd,
+        home,
+        _tmp_dir: tmp_dir,
     }
 }
