+++ boxlite-cli/tests/registry.rs
@@ -1,65 +1,64 @@
-use predicates::prelude::*;
+use boxlite_test_utils::TEST_REGISTRIES;
 
 mod common;
 
 #[test]
 fn test_run_with_custom_registry() {
-    let mut ctx = common::boxlite();
-    // This test relies on ghcr.io being available and hosting a 'hello-world' image.
-    // A more robust test would involve a local registry, but this is a good start.
-    ctx.cmd
-        .args(["run", "--rm", "--registry", "ghcr.io", "hello-world:latest"]);
-    ctx.cmd
-        .assert()
-        .success()
-        .stdout(predicate::str::contains("Hello from Docker!"));
+    let mut ctx = common::boxlite_bare();
+    let mirror = TEST_REGISTRIES[0];
+    ctx.cmd.arg("--registry").arg(mirror).args([
+        "run",
+        "--rm",
+        "alpine:latest",
+        "echo",
+        "custom registry works",
+    ]);
+    ctx.cmd.assert().success().stdout("custom registry works\n");
 }
 
 #[test]
 fn test_run_with_multiple_registries_fallback() {
-    let mut ctx = common::boxlite();
-    // First registry is invalid, should fall back to the second one (docker.io).
-    ctx.cmd.args([
-        "run",
-        "--rm",
-        "--registry",
-        "invalid.registry.that.does.not.exist",
-        "--registry",
-        "docker.io",
-        "alpine:latest",
-        "echo",
-        "hello from fallback",
-    ]);
+    let mut ctx = common::boxlite_bare();
+    ctx.cmd
+        .arg("--registry")
+        .arg("invalid.registry.that.does.not.exist")
+        .arg("--registry")
+        .arg(TEST_REGISTRIES[0])
+        .args([
+            "run",
+            "--rm",
+            "alpine:latest",
+            "echo",
+            "hello from fallback",
+        ]);
     ctx.cmd.assert().success().stdout("hello from fallback\n");
 }
 
 #[test]
 fn test_create_with_custom_registry() {
-    let mut ctx = common::boxlite();
+    let mut ctx = common::boxlite_bare();
+    let mirror = TEST_REGISTRIES[0];
     ctx.cmd
-        .args(["create", "--registry", "ghcr.io", "hello-world:latest"]);
+        .arg("--registry")
+        .arg(mirror)
+        .args(["create", "alpine:latest"]);
     let output = ctx.cmd.assert().success().get_output().clone();
     let box_id = String::from_utf8_lossy(&output.stdout).trim().to_string();
 
     assert!(!box_id.is_empty(), "Box ID should not be empty");
 
-    // Cleanup
     ctx.cleanup_box(&box_id);
 }
 
 #[test]
 fn test_run_fully_qualified_image_bypasses_registry() {
-    let mut ctx = common::boxlite();
-    // Provide an invalid registry, but a fully-qualified image name.
-    // The pull should succeed because the registry flag is ignored.
-    ctx.cmd.args([
-        "run",
-        "--rm",
-        "--registry",
-        "invalid.registry.that.does.not.exist",
-        "docker.io/library/alpine:latest",
-        "echo",
-        "fully qualified",
-    ]);
+    let mut ctx = common::boxlite_bare();
+    let qualified = format!("{}/library/alpine:latest", TEST_REGISTRIES[0]);
+    ctx.cmd
+        .arg("--registry")
+        .arg("invalid.registry.that.does.not.exist")
+        .arg("--registry")
+        .arg(TEST_REGISTRIES[0]) // needed for guest rootfs (debian:bookworm-slim)
+        .args(["run", "--rm", &qualified, "echo", "fully qualified"]);
     ctx.cmd.assert().success().stdout("fully qualified\n");
 }
