+++ boxlite/Cargo.toml
@@ -104,4 +104,5 @@ seccompiler = { version = "0.4", features = ["json"] }  # Compile JSON filters t
 bincode = "2.0"  # Serialize compiled BPF filters
 
 [dev-dependencies]
+boxlite-test-utils = { path = "../boxlite-test-utils" }
 proptest = "1.4"
